Scripts/requirements.txt
pip==21.0
PyYAML==5.4.1
setuptools==52.0.0
